+++ src/database/src/Model/Builder.php
@@ -1298,7 +1298,7 @@ class Builder
 
                 [$name, $constraints] = Str::contains($name, ':')
                     ? $this->createSelectWithConstraint($name)
-                    : [$name, function () {
+                    : [$name, static function () {
                     }];
             }
 
@@ -1321,7 +1321,7 @@ class Builder
      */
     protected function createSelectWithConstraint($name)
     {
-        return [explode(':', $name)[0], function ($query) use ($name) {
+        return [explode(':', $name)[0], static function ($query) use ($name) {
             $query->select(explode(',', explode(':', $name)[1]));
         }];
     }
@@ -1344,7 +1344,7 @@ class Builder
             $progress[] = $segment;
 
             if (! isset($results[$last = implode('.', $progress)])) {
-                $results[$last] = function () {
+                $results[$last] = static function () {
                 };
             }
         }